app/requirements.txt
torch
transformers
peft
fastapi
uvicorn
google-cloud-speech
google-cloud-texttospeech
bitsandbytes